% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_anno_apac.R
\name{AnnoAPAc}
\alias{AnnoAPAc}
\title{Annotate the APA clusters from a GTF file}
\usage{
AnnoAPAc(APAc_file, gtf_file, w_strand, add_chr = FALSE)
}
\arguments{
\item{APAc_file}{APA clusters file of one strand generated from ClusterAPA()}

\item{gtf_file}{a reference GTF file}

\item{w_strand}{which strand, "+" or "-"}

\item{add_chr}{whether to add character "chr" on the chr_name of the GTF (default FALSE)}
}
\value{
NULL. writes output to APA_clusters.out.anno in the out_dir
}
\description{
Annotate the APA clusters from a GTF file
}
\examples{
AnnoAPAc("output/positive_strand/APA_clusters.out", "ref/hg_19.gtf", "+")
AnnoAPAc("output/negative_strand/APA_clusters.out", "ref/hg_19.gtf", "-", TRUE)

}
