% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_cluster_apa.R
\name{ClusterAPA}
\alias{ClusterAPA}
\title{Cluster the two-dimensional coordinate files (RC and SA) to find the APA clusters}
\usage{
ClusterAPA(
  RC_prebed,
  SA_prebed,
  out_dir,
  n_endloc = 1e+05,
  blocksize = 100,
  lessmerge_n = 24,
  SA_weight = 10
)
}
\arguments{
\item{RC_prebed}{RC file of one strand generated from WriteEndlist()}

\item{SA_prebed}{SA file of one strand generated from WriteEndlist()}

\item{out_dir}{The directory where the output files are stored}

\item{n_endloc}{A approximate number of samples used in the first clustering, changes when processing (default and recommended 100000)}

\item{blocksize}{A approximate gap length and size of clusters, changes with the situation during clustering (default and recommended 100)}

\item{lessmerge_n}{The minimum size of APA clusters (default 24)}

\item{SA_weight}{The weight increase of SA file (default 10)}
}
\value{
NULL. writes output to APA_clusters.out in the out_dir
}
\description{
Cluster the two-dimensional coordinate files (RC and SA) to find the APA clusters
}
\examples{
ClusterAPA("output/negative_strand/sa.prebed", "output/negative_strand/sa.prebed", "output/negative_strand/")
ClusterAPA("output/positive_strand/sa.prebed", "output/positive_strand/sa.prebed", "output/positive_strand/")
}
