% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_count_cb.R
\name{CountCB}
\alias{CountCB}
\title{Create a APA_cluster-Cell_Barcode count matrix for Seurart analysis}
\usage{
CountCB(cluster_p, cluster_n, in_bamfile, cb_file, chr_info, out_dir)
}
\arguments{
\item{cluster_p}{Annotated APA clusters file of the positive strand, generated by AnnoAPAc()}

\item{cluster_n}{Annotated APA clusters file of the negative strand, generated by AnnoAPAc()}

\item{in_bamfile}{A indexed BAM file, and the index files (.bai) must be in the same directory of the indexed BAM}

\item{cb_file}{A cell barcodes list, can be generated by WriteEndlist() or cellranger}

\item{chr_info}{A data.table that contains the chrinfo, can be generated by ExtractChrinfo()}

\item{out_dir}{The directory where a directory named "matrix/" is created and the output files are stored}
}
\value{
NULL. writes output in the directory "matrix/" of the out_dir
}
\description{
Create a APA_cluster-Cell_Barcode count matrix for Seurart analysis
}
\examples{
CountCB("output/positive_strand/APA_clusters.out.anno",
    "output/negative_strand/APA_clusters.out.anno","input/test.sorted.bam", "output/_CBlist.txt",
    ExtractChrinfo("input/test.sorted.bam"), "output/")

}
