% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_extract_chrinfo.R
\name{ExtractChrinfo}
\alias{ExtractChrinfo}
\title{Extracting the chromosomes information from the indexed BAM file, including names and length.}
\usage{
ExtractChrinfo(in_bamfile)
}
\arguments{
\item{in_bamfile}{A indexed BAM file, and the index files (.bai) must be in the same directory of the indexed BAM}
}
\value{
A data.table with chromosomes names and length
}
\description{
Extracting the chromosomes information from the indexed BAM file, including names and length.
}
\examples{
chr_info <- ExtractChrinfo("input/test.sorted.bam")

}
