% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_write_endlist.R
\name{WriteEndlist}
\alias{WriteEndlist}
\title{Generate two-dimensional coordinate files of each strand for clustering}
\usage{
WriteEndlist(in_bamfile, in_nSBfile, out_dir, chr_info, RCn = 3)
}
\arguments{
\item{in_bamfile}{A indexed BAM file, and the index files (.bai) must be in the same directory of the indexed BAM}

\item{in_nSBfile}{A indexed soft-clipped BAM file, generated and indexed by samtools script}

\item{out_dir}{The directory where directories of two strands are created and the output files are stored}

\item{chr_info}{A data.table that contains the chrinfo, can be generated by ExtractChrinfo()}

\item{RCn}{The cut-off value that decides the minimum weight of output when processing the in_bamfile (default 3)}
}
\value{
NULL. writes output to file, SA files are generated from in_nSBfile and RC files are from in_bamfile.
}
\description{
Generate two-dimensional coordinate files of each strand for clustering
}
\examples{
WriteEndlist("input/test.sorted.bam", "input/3S_test.sorted.bam", "output/", ExtractChrinfo("input/test.sorted.bam"))
WriteEndlist("input/3S_test.sorted.bam", "input/3S_test.sorted.bam", "output/", ExtractChrinfo("input/3S_test.sorted.bam"))
}
